pyproject.toml
[project]
name = "ruleslawyerbot"
version = "0.1.0"
description = "🎲 Telegram bot that acts as a board game rules referee using OpenAI Agents SDK with multi-stage conversational pipeline, multilingual PDF search, and inline UI for answering game rules questions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.8.1",
    "openai-agents>=0.6.1",
    "pypdf>=6.3.0",
    "python-telegram-bot>=22.5",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.8",
    "pydantic-settings>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.19.1",
    "logfire>=3.0.0",
]

[tool.uv]
package = true
